Improve unit testing and add per-type functions

+++ src/main/java/com/revolvingmadness/mclang/types/FloatType.java
@@ -2,11 +2,11 @@ package com.revolvingmadness.mclang.types;
 
 public class FloatType extends NumberType {
     public FloatType(Number value) {
-        super(value.floatValue());
+        super(value.doubleValue());
     }
 
     @Override
     public String toString() {
-        return Float.toString(value.floatValue());
+        return Double.toString(value.doubleValue());
     }
 }
